Modified to use top-level  instead of extending gcs API
